requirements.txt
certifi==2022.9.14
charset-normalizer==2.1.1
flake8==5.0.4
idna==3.4
importlib-metadata==4.2.0
mccabe==0.7.0
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==0.21.0
requests==2.28.1
typing_extensions==4.3.0
urllib3==1.26.12
vk-api==11.9.9
zipp==3.8.1
